requirements.txt
Django==4.0.1
django-environ==0.9.0
sqlparse==0.4.3
psycopg2-binary==2.9.5
djangorestframework==3.14.0
django-cors-headers==3.13.0
django-filter==22.1
social-auth-app-django==5.0.0
social-auth-core==4.3.0
requests-oauthlib==1.3.1
drf-spectacular==0.25.1
marshmallow==3.19.0
marshmallow-dataclass==8.5.10
requests==2.28.1
pytest-django==4.5.2
pytest-factoryboy==2.5.1
pytest~=7.2.0